test: Fix flaky test
Split the `await_cross_thread_guard` tests into separate files. This should hopefully fix the flakiness in #1112 by isolating the tests.

Fixes [#1112](https://github.com/getsentry/sentry-rust/issues/1112)
Fixes [RUST-210](https://linear.app/getsentry/issue/RUST-210/flaky-test-future-cross-thread-info-span)
